+++ tests/testthat.R
@@ -1,4 +1,3 @@
-library(httptest)
+library(testthat)
 library(echor)
-
 test_check("echor")